requirements.txt
paho-mqtt==1.6.1
PyYAML==6.0.1
setuptools~=65.5.1
requests~=2.31.0